first cut at refactoring with own dpk_lang_id name space
