+++ transforms/language/lang_id/Dockerfile.python
@@ -24,10 +24,9 @@ RUN apt update && apt install gcc g++ -y
 RUN mkdir -p /home/dpk/.cache/huggingface/hub && chmod -R 777 /home/dpk/.cache/huggingface/hub
 USER dpk
 
-COPY --chown=dpk:root src/ src/
-COPY --chown=dpk:root pyproject.toml pyproject.toml 
+COPY --chown=dpk:root dpk_lang_id/ dpk_lang_id/
 COPY --chown=dpk:root requirements.txt requirements.txt
-RUN pip install --no-cache-dir -e .
+RUN pip install --no-cache-dir -r requirements.txt
 
 # clean up apt
 USER root 
@@ -36,18 +35,6 @@ RUN apt-get remove gcc g++ -y \
     && rm -rf /var/cache/apt/archives/* /var/lib/apt/lists/*
 USER dpk
 
-#COPY requirements.txt requirements.txt
-#RUN pip install --no-cache-dir -r  requirements.txt
-
-# copy the main() entry point to the image 
-COPY ./src/lang_id_transform_python.py .
-
-# copy some of the samples in
-COPY ./src/lang_id_local.py local/
-
-# copy test
-COPY test/ test/
-COPY test-data/ test-data/
 
 # Set environment
 ENV PYTHONPATH /home/dpk
