+++ transforms/language/lang_id/dpk_lang_id/local_python.py
@@ -13,18 +13,16 @@
 import os
 import sys
 
+from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.utils import ParamsUtils
-from data_processing_ray.runtime.ray import RayTransformLauncher
-from lang_id_transform import (
+from dpk_lang_id.lang_models import KIND_FASTTEXT
+from dpk_lang_id.transform import (
     content_column_name_cli_param,
     model_credential_cli_param,
     model_kind_cli_param,
     model_url_cli_param,
-    output_lang_column_name_cli_param,
-    output_score_column_name_cli_param,
 )
-from lang_id_transform_ray import LangIdentificationRayTransformConfiguration
-from lang_models import KIND_FASTTEXT
+from dpk_lang_id.transform_python import LangIdentificationPythonTransformConfiguration
 
 
 # create parameters
@@ -34,32 +32,24 @@ local_conf = {
     "input_folder": input_folder,
     "output_folder": output_folder,
 }
-worker_options = {"num_cpus": 0.8}
 code_location = {"github": "github", "commit_hash": "12345", "path": "path"}
 params = {
-    # where to run
-    "run_locally": True,
     # Data access. Only required parameters are specified
     "data_local_config": ParamsUtils.convert_to_ast(local_conf),
-    # orchestrator
-    "runtime_worker_options": ParamsUtils.convert_to_ast(worker_options),
-    "runtime_num_workers": 3,
+    # execution info
     "runtime_pipeline_id": "pipeline_id",
     "runtime_job_id": "job_id",
-    "runtime_creation_delay": 0,
     "runtime_code_location": ParamsUtils.convert_to_ast(code_location),
     # lang_id params
     model_credential_cli_param: "PUT YOUR OWN HUGGINGFACE CREDENTIAL",
     model_kind_cli_param: KIND_FASTTEXT,
     model_url_cli_param: "facebook/fasttext-language-identification",
     content_column_name_cli_param: "text",
-    output_lang_column_name_cli_param: "ft_lang",
-    output_score_column_name_cli_param: "ft_score",
 }
 if __name__ == "__main__":
     # Set the simulated command line args
     sys.argv = ParamsUtils.dict_to_req(d=params)
     # create launcher
-    launcher = RayTransformLauncher(LangIdentificationRayTransformConfiguration())
+    launcher = PythonTransformLauncher(runtime_config=LangIdentificationPythonTransformConfiguration())
     # Launch the ray actor(s) to process the input
     launcher.launch()
