+++ transforms/language/lang_id/dpk_lang_id/ray/s3.py
@@ -13,43 +13,59 @@
 import os
 import sys
 
-from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.utils import ParamsUtils
-from lang_id_transform import (
+from data_processing_ray.runtime.ray import RayTransformLauncher
+from dpk_lang_id.lang_models import KIND_FASTTEXT
+from dpk_lang_id.ray.transform import LangIdentificationRayTransformConfiguration
+from dpk_lang_id.transform import (
     content_column_name_cli_param,
     model_credential_cli_param,
     model_kind_cli_param,
     model_url_cli_param,
+    output_lang_column_name_cli_param,
+    output_score_column_name_cli_param,
 )
-from lang_id_transform_python import LangIdentificationPythonTransformConfiguration
-from lang_models import KIND_FASTTEXT
 
 
+print(os.environ)
+# create launcher
+launcher = RayTransformLauncher(LangIdentificationRayTransformConfiguration())
 # create parameters
-input_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "test-data", "input"))
-output_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "output"))
-local_conf = {
-    "input_folder": input_folder,
-    "output_folder": output_folder,
+s3_cred = {
+    "access_key": "localminioaccesskey",
+    "secret_key": "localminiosecretkey",
+    "url": "http://localhost:9000",
 }
+s3_conf = {
+    "input_folder": "test/lang_id/input",
+    "output_folder": "test/lang_id/output",
+}
+worker_options = {"num_cpus": 0.8}
 code_location = {"github": "github", "commit_hash": "12345", "path": "path"}
 params = {
+    # where to run
+    "run_locally": True,
     # Data access. Only required parameters are specified
-    "data_local_config": ParamsUtils.convert_to_ast(local_conf),
-    # execution info
+    "data_s3_cred": ParamsUtils.convert_to_ast(s3_cred),
+    "data_s3_config": ParamsUtils.convert_to_ast(s3_conf),
+    # orchestrator
+    "runtime_worker_options": ParamsUtils.convert_to_ast(worker_options),
+    "runtime_num_workers": 3,
     "runtime_pipeline_id": "pipeline_id",
     "runtime_job_id": "job_id",
+    "runtime_creation_delay": 0,
     "runtime_code_location": ParamsUtils.convert_to_ast(code_location),
     # lang_id params
     model_credential_cli_param: "PUT YOUR OWN HUGGINGFACE CREDENTIAL",
     model_kind_cli_param: KIND_FASTTEXT,
     model_url_cli_param: "facebook/fasttext-language-identification",
     content_column_name_cli_param: "text",
+    output_lang_column_name_cli_param: "ft_lang",
+    output_score_column_name_cli_param: "ft_score",
 }
-if __name__ == "__main__":
-    # Set the simulated command line args
-    sys.argv = ParamsUtils.dict_to_req(d=params)
-    # create launcher
-    launcher = PythonTransformLauncher(runtime_config=LangIdentificationPythonTransformConfiguration())
-    # Launch the ray actor(s) to process the input
-    launcher.launch()
+sys.argv = ParamsUtils.dict_to_req(d=params)
+# for arg in sys.argv:
+#     print(arg)
+
+# launch
+launcher.launch()
