+++ transforms/language/lang_id/dpk_lang_id/transform_python.py
@@ -10,21 +10,20 @@
 # limitations under the License.
 ################################################################################
 
-import pyarrow as pa
-from data_processing.utils import get_logger
-from data_processing_ray.runtime.ray import RayTransformLauncher
-from data_processing_ray.runtime.ray.runtime_configuration import (
-    RayTransformRuntimeConfiguration,
+from data_processing.runtime.pure_python import PythonTransformLauncher
+from data_processing.runtime.pure_python.runtime_configuration import (
+    PythonTransformRuntimeConfiguration,
 )
-from lang_id_transform import LangIdentificationTransformConfiguration
+from data_processing.utils import get_logger
+from dpk_lang_id.transform import LangIdentificationTransformConfiguration
 
 
 logger = get_logger(__name__)
 
 
-class LangIdentificationRayTransformConfiguration(RayTransformRuntimeConfiguration):
+class LangIdentificationPythonTransformConfiguration(PythonTransformRuntimeConfiguration):
     """
-    Implements the RayTransformConfiguration for Language Identification as required by the RayTransformLauncher.
+    Implements the PythonTransformConfiguration for Language Identification as required by the PythonTransformLauncher.
     Language Identification does not use a RayRuntime class so the superclass only needs the base
     python-only configuration.
     """
@@ -38,6 +37,6 @@ class LangIdentificationRayTransformConfiguration(RayTransformRuntimeConfigurati
 
 
 if __name__ == "__main__":
-    launcher = RayTransformLauncher(LangIdentificationRayTransformConfiguration())
+    launcher = PythonTransformLauncher(LangIdentificationPythonTransformConfiguration())
     logger.info("Launching lang_id transform")
     launcher.launch()
