--- transforms/language/lang_id/python/.dockerignore
@@ -1,1 +1,0 @@
-venv/
