--- transforms/language/lang_id/ray/.dockerignore
@@ -1,1 +1,0 @@
-venv/
